Fix display of CNAME/A records in log file.
print_dns_result() will show the number of records
found _and_ put a heading for this type of record
into the logfile.

So we can't get both right (short of changing
print_dns_result(), what I don't like to do
just now): either the count will be too high,
as we can't subtract wrong records from it, or
the "heading" will follow after the list of
records in the log. So we'll have to live with
the former, until print_dns_result() can be
fixed.

Signed-off-by: Thomas Hochstein <thh@inter.net>